+++ koku/api/iam/models.py
@@ -91,18 +91,6 @@ class UserPreference(models.Model):
                                                                      self.preference)
 
 
-class CostUsageReportStatus(models.Model):
-    """Information on the state of the cost usage report."""
-
-    provider = models.ForeignKey('Provider', null=True,
-                                 on_delete=models.CASCADE)
-    report_name = models.CharField(max_length=128, null=False, unique=True)
-    cursor_position = models.PositiveIntegerField()
-    last_completed_datetime = models.DateTimeField(null=True)
-    last_started_datetime = models.DateTimeField(null=True)
-    etag = models.CharField(max_length=64, null=True)
-
-
 class Tenant(TenantMixin):
     """The model used to create a tenant schema."""
