+++ koku/api/models.py
@@ -17,8 +17,22 @@
 """API models for import organization."""
 # flake8: noqa
 # pylint: disable=unused-import
+from django.db import models
+
 from api.status.models import Status
-from api.iam.models import Customer, ResetToken, Tenant, User, UserPreference, CostUsageReportStatus
+from api.iam.models import Customer, ResetToken, Tenant, User, UserPreference
 from api.provider.models import (Provider,
                                  ProviderAuthentication,
                                  ProviderBillingSource)
+
+
+class CostUsageReportStatus(models.Model):
+    """Information on the state of the cost usage report."""
+
+    provider = models.ForeignKey('Provider', null=True,
+                                 on_delete=models.CASCADE)
+    report_name = models.CharField(max_length=128, null=False, unique=True)
+    cursor_position = models.PositiveIntegerField()
+    last_completed_datetime = models.DateTimeField(null=True)
+    last_started_datetime = models.DateTimeField(null=True)
+    etag = models.CharField(max_length=64, null=True)
